Directory: Use the theme's page template as the directory template.
On the Social Paper directory page, instead of the theme's archive
template, we now use the theme's page template for the template wrapper as
it is more generic-looking.  This also fits in with BuddyPress directory
pages as BP uses the page template.

This commit also restructures our content directory template part to
be more like a page template part.  This is so our template will fit in
seamlessly with the current theme.

See #5.