FEE: Do not add 'new' page to the browser's history.
If you click on the 'New Paper' link, you will land on the paper writing
interface.  However, if you click on the browser's 'Back' button, you will
hit the 'new' page and get redirected to the paper writing interface again.

This is because for SP, we use the 'new' page as a redirector.

To allow the 'Back' button to work properly, we have to override FEE's
adminbar JS and use a customized version of the JS.

See #8.