Don't let FEE force `comment_status` to false for new papers.
FEE falls back on 'default_comment_status' for new documents. This value is
enforced by the fact that the auto-draft created during `ajax_new` becomes
the `postOnServer`, which is used to populate the `comment_status` in
`ajax_post`. I think it's a bug in FEE that it's so strict about this. In any
case, it's a problem for SP, because it means that new papers have comments
disabled if 'default_comment_status' is also off.

Fixes #26.