Basic notification support for "someone commented on a thread you're a member of".
This is kind of a mess. What counts as a "thread" is not really clear, so for
now users are notified if a comment is left on any post where they've
commented. This is overly broad, and we can look for a better definition of
"thread" in the future.

See #22.