Introduce a custom `isDirty` method.
Instead of comparing content, we set an `isDirty` flag whenever a change is
detected. The flag gets set back to `false` on `fee-after-save`.

This technique can result in some false positives, as when you change content
but then undo that change. But it has a couple of benefits that outweigh this
downside:

* It warns the user about unsaved changes on the Settings panel, which FEE doesn't know about.
* It allows us to modify the content of the editor at will, without worrying about FEE's `isDirty` checks. See #77, #71, #42.