Always show 'Responses' header in comment section.
Even when the paper has no comments.

We're painting with a broad brush here, because the `have_comments()` check
counts all comments on a post, including those that are displayed inline
rather than at the end of the paper. Down the line, we might make the language
here more sensitive to whether there are comments that appear at the end of the
paper.

Fixes #93.