Persistent caching for potential paper readers.
The caching is very coarse-grained at the moment: potential readers for a given
dyad of logged-in user + paper_id are cached in a single array, and the cache
is liberally busted using a 'last_changed' incrementor that busts all
'potential_readers' caches. Much of this caching would not be necessary if
BP itself had better caching for friend and group member queries. As such,
it's likely that the current caching schema will be yanked out at some point.